bus/hp_dio/98603b: convert to anonymous namespace (nw)
